refactor: change server structure and correct imports
