+++ package.json
@@ -57,6 +57,7 @@
     "react-hook-form": "^7.50.1",
     "react-resizable-panels": "^2.0.9",
     "resend": "^3.2.0",
+    "server-only": "^0.0.1",
     "sharp": "^0.33.3",
     "tailwind-merge": "^2.2.1",
     "tailwind-scrollbar": "^3.1.0",
