+++ src/components/Auth/LoginForm.tsx
@@ -17,8 +17,8 @@ import { Input } from '../ui/input';
 import { login } from '@/actions/login';
 import { useEffect, useState, useTransition } from 'react';
 import { loginFormSchema } from '@/schemas';
-import GoogleLogo from '@assets/third-part-login/Google.png';
-import GithubLogo from '@assets/third-part-login/GitHub.svg';
+import GoogleLogo from '@/assets/third-part-login/Google.png';
+import GithubLogo from '@/assets/third-part-login/GitHub.svg';
 import { useToast } from '../ui/use-toast';
 import { X } from 'lucide-react';
 import { useSearchParams } from 'next/navigation';
