+++ src/server/actions/new-password.ts
@@ -1,11 +1,11 @@
 'use server';
 
-import { getPasswordResetTokenByToken } from '@/data/password-reset-token';
-import { getUserByEmail } from '@/data/user';
+import { getPasswordResetTokenByToken } from '@/server/queries/password-reset-token';
+import { getUserByEmail } from '@/server/queries/user';
 import { newPasswordSchema } from '@/schemas';
 import { z } from 'zod';
 import bycrpt from 'bcryptjs';
-import { db } from '@/db';
+import { db } from '@/server/db';
 
 export async function newPassword(
   values: z.infer<typeof newPasswordSchema>,
