+++ src/server/actions/register.ts
@@ -3,8 +3,8 @@
 import { registerFormSchema } from '@/schemas';
 import { z } from 'zod';
 import bcrypt from 'bcryptjs';
-import { db } from '@/db';
-import { getUserByEmail } from '@/data/user';
+import { db } from '@/server/db';
+import { getUserByEmail } from '@/server/queries/user';
 import { signIn } from '@/auth';
 import { DEFAULT_REDIRECT } from '@/routes';
 import { AuthError } from 'next-auth';
