+++ tsconfig.json
@@ -19,7 +19,10 @@
     ],
     "paths": {
       "@/*": ["./src/*"],
-      "@assets/*": ["./public/assets/*"]
+      "@/assets/*": ["./public/assets/*"],
+      "@/server/*": ["./src/server/*"],
+      "@/queries/*": ["./src/server/queries/*"],
+      "@/actions/*": ["./src/server/actions/*"]
     }
   },
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
